Extra check in union_find list, real r1c analysis
Without extra check allows for double growth of cluster C0 in support 0 
bucket, e.g. [C0, C0]. This is due to that in list method a cluster is 
now directly placed after growth, but the bucket is growth before 
placing. 

Round1clusters used to be clusters of error qubit locations (unknown 
information), rather than clusters of grown (edge.support == 2) edges. 
New analysis is setup for this.